+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Unreleased
 
+Added:
+
+- Static health check
+
 ## 0.1.0 (December 20, 2020)
 
 - Initial release